fix: refactor profile component
