+++ src/components/app-routes/app-routes.tsx
@@ -9,8 +9,8 @@ import { Register } from '@/pages/register'
 import { ForgotPassword } from '@/pages/forgot-password'
 import { ResetPassword } from '@/pages/reset-password'
 import { Profile } from '@/pages/profile'
-import { ProfileInfo } from '@/pages/profile/profile-order'
-import { ProfileOrder } from '@/pages/profile/profile-info'
+import { ProfileInfo } from '@/pages/profile/profile-info'
+import { ProfileOrder } from '@/pages/profile/profile-order'
 import { IngredientDetails } from '@/components/burger-ingredients/ingredient-details'
 import { ProtectedRouteOnlyAuth, ProtectedRouteOnlyUnAuth } from '@/components/protected-route'
 import { AppRoutesProps } from './types/app-routes-props'
