+++ src/components/form/form-button/form-button.tsx
@@ -1,5 +1,6 @@
 import { FC, useState } from 'react'
 import { Button } from '@ya.praktikum/react-developer-burger-ui-components'
+
 import { FormButtonProps } from './types/form-button-props'
 
 const FormButton: FC<FormButtonProps> = ({ onClick, buttonText }) => {
