+++ src/pages/profile/profile-info/index.ts
@@ -1,1 +1,1 @@
-export { default as ProfileOrder } from './profile-order'
+export { default as ProfileInfo } from './profile-info'
