+++ src/pages/profile/profile-info/profile-info.tsx
@@ -1,7 +1,7 @@
 import { UpdateUserDTO } from '@/utils/types'
-import { FormUpdate } from '@/components/form/form-update'
 import { useAppDispatch } from '@/services/store'
 import { update } from '@/services/user/actions'
+import { FormUpdate } from '@/components/form/form-update'
 
 const ProfileInfo = () => {
   const dispatch = useAppDispatch()
