+++ src/scss/layout/_layout.scss
@@ -31,7 +31,7 @@
 
   &_profile {
     .profile__text {
-      opacity: .4;
+      opacity: 0.4;
     }
   }
 }
